WIP #392: rename schools.py to schools_2021.py for testing vz_google_sheets_2021.py
