+++ dags/vz_google_sheets_2021.py
@@ -42,7 +42,7 @@ def task_fail_slack_alert(context):
 #to read the python script for pulling data from google sheet and putting it into tables in postgres
 try:
   sys.path.append('/etc/airflow/data_scripts/vision_zero/')
-  from schools import pull_from_sheet
+  from schools_2021 import pull_from_sheet
 except:
   raise ImportError("Cannot import functions to pull school safety zone list")
